- création d'un module d'interface utilisateur : CosmoGUI la GUI est contrôlable à la souris, au clavier, et au joystick - refonte complète des menus (utilisation de boutons, sliders, checboxes...) - plus d'options dans le menu "Son et musique"
fix issue #16

+++ src/core/LevelManager.cpp
@@ -168,7 +168,7 @@ LevelManager::Error LevelManager::ParseFile(const char* file)
 		node = node->NextSibling();
 	}
 #ifdef DEBUG
-	printf("%d levels found\n", levels_.size());
+	printf("%d levels found\n", (int) levels_.size());
 #endif
 	return SUCCESS;
 }
